+++ src/main/java/services/UserService.java
@@ -1,4 +1,4 @@
-package service;
+package services;
 
 import dto.RegistrationDto;
